+++ roqoqo/src/operations/mod.rs
@@ -61,9 +61,7 @@ pub use three_qubit_gate_operations::*;
 mod bosonic_operations;
 pub use bosonic_operations::*;
 /// Collection of roqoqo spin-boson operations.
-#[cfg(feature = "unstable_spin_boson_operations")]
 mod spin_boson_operations;
-#[cfg(feature = "unstable_spin_boson_operations")]
 pub use spin_boson_operations::*;
 /// Collection of roqoqo analog gate operations
 #[cfg(feature = "unstable_analog_operations")]