+++ CHANGELOG.md
@@ -2,9 +2,13 @@
 
 This changelog track changes to the qoqo project starting at version v0.5.0
 
-## 1.11
+## 1.11.0
 
-### Added in 1.11
+### Added in 1.11.0-alpha.1
+
+* Made operations in `unstable_spin_boson_operations` stable
+
+### Added in 1.11.0-alpha.0
 
 * Unstable feature `unstable_analog_operations` adding new analog operations: `ApplyConstantSpinHamiltonian`, `ApplyTimeDependentSpinHamiltonian`
 * Added `SingleQubitOverrotationDescription`
@@ -16,6 +20,7 @@ This changelog track changes to the qoqo project starting at version v0.5.0
 ### Added in 1.10.0
 
 * Unstable feature `unstable_spin_boson_operations` adding new spin-boson operations: QuantumRabi, LongitudinalCoupling, JaynesCummings
+* Added three new operations to `unstable_spin_boson_operations`: SingleExcitationLoad, SingleExcitationStore, CZQubitResonator
 
 ### Fixed in 1.9.2